+++ app/scripts/lib/lforms-data.js
@@ -77,6 +77,9 @@
     // a delimiter used in code path and id path
     PATH_DELIMITER : "/",
 
+    // whether the form data contains saved user data
+    hasSavedData : false,
+
     // repeatable question items derived from items
     _repeatableItems : {},
 
@@ -347,10 +350,10 @@
     /**
      * Calculate internal data from the raw form definition data,
      * including:
-     * structural data, (TBD: unless they are already included (when hasUserData == true) ):
+     * structural data:
      *    _id, _idPath, _codePath
      * data for widget control and/or performance improvement:
-     *    _displayLevel_,
+     *    _displayLevel_
      * @private
      */
     _initializeInternalData: function() {
@@ -380,7 +383,7 @@
       this.Navigation.setupNavigationMap(this);
 
       // create auto-completer options and assign field default values
-      this._setUpDefaultsAndAutocomp();
+      this._setUpAnswerAndUnitAutoComp(this.itemList);
 
       // set up a mapping from controlling items to controlled items
       // for skip logic, data controls and formulas
@@ -520,7 +523,7 @@
       this.Navigation.setupNavigationMap(this);
 
       // create auto-completer options
-      this._setUpDefaultsAndAutocomp();
+      this._setUpAnswerAndUnitAutoComp(this.itemList);
 
       // set up a mapping from controlling items to controlled items
       // for skip logic, data controls and formulas
@@ -883,6 +886,7 @@
           if (item.value && (item.dataType === this._CONSTANTS.DATA_TYPE.DT || item.dataType === this._CONSTANTS.DATA_TYPE.DTM)) {
               item.value = LForms.Util.stringToDate(item.value);
           }
+
         }
       }
     },
@@ -934,9 +938,27 @@
           this._mergeTwoArrays(this.templateOptions.columnHeaders, columnHeaders);
         }
 
-        // if there is a new formHeaderItems array, set up autocomplete options
-        if (newOptions.formHeaderItems)
-          this._setUpDefaultsAndAutocomp(true);
+        // if there is a formHeaderItems array, set up autocomplete options
+        if (this.templateOptions.formHeaderItems) {
+          for (var i=0, iLen=this.templateOptions.formHeaderItems.length; i<iLen; i++) {
+            var item = this.templateOptions.formHeaderItems[i];
+            if (item.dataType === this._CONSTANTS.DATA_TYPE.CWE ||
+                item.dataType === this._CONSTANTS.DATA_TYPE.CNE) {
+              this._updateAutocompOptions(item);
+              this._resetItemValueWithModifiedAnswers(item);
+            }
+            // if this is not a saved form with user data, and
+            // there is a default value, and
+            // there is no embedded data
+            else if (!this.hasSavedData && item.defaultAnswer && !item.value) {
+              item.value = item.defaultAnswer;
+            }
+            this._updateUnitAutocompOptions(item);
+          }
+
+        }
+
+
       }
     },
 
@@ -1022,7 +1044,16 @@
         this._setModifiedAnswers(item); // sets item._modifiedAnswers
 
         // reset item.value with modified answers if the item has a value (or an array of values)
-        this._resetItemValueWithModifiedAnswers(item)
+        if (item.dataType === this._CONSTANTS.DATA_TYPE.CWE ||
+            item.dataType === this._CONSTANTS.DATA_TYPE.CNE) {
+          this._resetItemValueWithModifiedAnswers(item);
+        }
+        // if this is not a saved form with user data, and
+        // there is a default value, and
+        // there is no embedded data
+        else if (!this.hasSavedData && item.defaultAnswer && !item.value) {
+          item.value = item.defaultAnswer;
+        }
 
         // normalize unit value if there is one, needed by calculationMethod
         if (item.unit && !item.unit.text) {
@@ -1319,6 +1350,16 @@
       // get the form data
       var formData = this.getUserData(false, noEmptyValue, noHiddenItem, keepIdPath, keepCodePath);
 
+      // check if there is user data
+      var hasSavedData = false;
+      for (var i=0, iLen=this.itemList.length; i<iLen; i++) {
+        var item = this.itemList[i];
+        if (!LForms.Util.isItemValueEmpty(item)) {
+          hasSavedData = true;
+          break;
+        }
+      }
+
       var defData = {
         PATH_DELIMITER: this.PATH_DELIMITER,
         code: this.code,
@@ -1332,8 +1373,13 @@
         items: formData.itemsData,
         templateOptions: angular.copy(this.templateOptions)
       };
+
+      if (hasSavedData) {
+        defData.hasSavedData = true;
+      }
+
       // reset obr fields
-      defData.templateOptions.formHeaderItems = formData.templateData;
+      defData.templateOptions.formHeaderItems = angular.copy(formData.templateData);
 
       return defData;
     },
@@ -2355,26 +2401,17 @@
 
 
     /**
-     * Set up autocomplete options for each items
-     * @param templateOptionsOnly (default false) set to true if only the
-     *  templateOptions items need processing.
+     * Set up autocomplete options for each item
+     * @param items a list items of the form or in the templateOptions.
      */
-    _setUpDefaultsAndAutocomp: function(templateOptionsOnly) {
-      var itemList;
-      var itemLists = [this.templateOptions.formHeaderItems];
-      if (!templateOptionsOnly)
-        itemLists.push(this.itemList);
-      for (var j=0, jLen=itemLists.length; j<jLen && (itemList = itemLists[j]); ++j) {
-        for (var i=0, iLen=itemList.length; i<iLen; i++) {
-          var item = itemList[i];
-          if (item.dataType === this._CONSTANTS.DATA_TYPE.CWE ||
-              item.dataType === this._CONSTANTS.DATA_TYPE.CNE) {
-            this._updateAutocompOptions(item);
-          }
-          else if (item.defaultAnswer && !item.value) // && not a list
-            item.value = item.defaultAnswer;
-          this._updateUnitAutocompOptions(item);
+    _setUpAnswerAndUnitAutoComp: function(items) {
+      for (var i=0, iLen=items.length; i<iLen; i++) {
+        var item = items[i];
+        if (item.dataType === this._CONSTANTS.DATA_TYPE.CWE ||
+            item.dataType === this._CONSTANTS.DATA_TYPE.CNE) {
+          this._updateAutocompOptions(item);
         }
+        this._updateUnitAutocompOptions(item);
       }
     },
 
@@ -2530,7 +2567,8 @@
         // default answer and item.value could be a string value, if it is a not-on-list value for CWE types
         var modifiedValue = null;
         // item.value has the priority over item.defaultAnswer
-        var answerValue = item.value || item.defaultAnswer;
+        // if this is a saved form with user data, default answers are not to be used.
+        var answerValue = this.hasSavedData ? item.value : item.value || item.defaultAnswer;
         if (answerValue) {
           modifiedValue = [];
           // could be an array of multiple default values or a single value
@@ -2803,9 +2841,10 @@
             options.itemToHeading = itemToHeading;
           }
 
-          // If there isn't already a default value set (handled elsewhere), and
+          // If this is not a saved form with user data, and
+          // there isn't already a default value set (handled elsewhere), and
           // there is just one item in the list, use that as the default value.
-          if (!options.defaultValue && options.listItems.length === 1)
+          if (!this.hasSavedData && !options.defaultValue && options.listItems.length === 1)
             options.defaultValue = options.listItems[0];
         }
         item._autocompOptions = options;