+++ test/protractor/spec/lforms_fhirdata.spec.js
@@ -102,6 +102,7 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
               expect(fhirData.resourceType).toBe("DiagnosticReport");
               expect(fhirData.result.length).toBe(0);
               expect(fhirData.contained).toEqual([]);
+
               // #2 has some values
               // ST, repeating
               ff.name.sendKeys("name 1");
@@ -267,6 +268,7 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
               expect(fhirData.entry.length).toBe(1);
               expect(fhirData.entry[0].resource.resourceType).toBe("DiagnosticReport");
               expect(fhirData.entry[0].resource.result).toEqual([]);
+
               // #2 has some values
               // ST, repeating
               ff.name.sendKeys("name 1");
@@ -311,6 +313,14 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
               ff.ageAtDiag2.sendKeys(protractor.Key.ARROW_DOWN);
               ff.ageAtDiag2.sendKeys(protractor.Key.TAB);
 
+              // // remove the default values on the 2 items
+              // expect(ff.related.getAttribute('value')).toEqual('No');
+              // ff.related.clear();
+              // expect(ff.related.getAttribute('value')).toEqual('');
+              //
+              // expect(ff.mockedHeight.getAttribute('value')).toEqual('72');
+              // ff.mockedHeight.clear();
+              // expect(ff.mockedHeight.getAttribute('value')).toEqual('');
 
               getFHIRResource("DiagnosticReport", fhirVersion,
                   {bundleType: "collection"}).then(function(callbackData) {
@@ -617,7 +627,6 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
 
           it('should merge all DiagnosticReport (contained) data back into the form', function() {
 
-            tp.openUSSGFHTVertical();
             tp.setFHIRVersion(fhirVersion);
 
             element(by.id("merge-dr")).click();
@@ -653,7 +662,6 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
 
           it('should merge all DiagnosticReport (Bundle) data back into the form', function() {
 
-            tp.openUSSGFHTVertical();
             tp.setFHIRVersion(fhirVersion);
 
             element(by.id("merge-bundle-dr")).click();
@@ -675,6 +683,40 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
             expect(ff.disease.getAttribute('value')).toBe("Hypertension");
             expect(ff.ageAtDiag.getAttribute('value')).toBe("Newborn");
           });
+
+
+          it('should merge all DiagnosticReport (contained) data back into the form without setting default values', function() {
+
+            //tp.openUSSGFHTVertical();
+            tp.setFHIRVersion(fhirVersion);
+
+            element(by.id("merge-dr-default-values")).click();
+            browser.waitForAngular();
+
+            var intField = element(by.id('/intField/1')),
+                decField = element(by.id('/decField/1')),
+                strField = element(by.id('/strField/1')),
+                dateField = element(by.id('/dateField/1')),
+                listField = element(by.id('/ansCodeDefault/1'));
+
+            browser.wait(function() {
+              try {
+                return intField.isDisplayed(); // sometimes results in a "stale reference" error
+              }
+              catch (e) {
+                // Try to refresh the element
+                intField = element(by.id('/intField/1'));
+              }
+            }, tp.WAIT_TIMEOUT_1);
+
+            expect(intField.getAttribute('value')).toBe('24'); // it is a value in dr
+            expect(decField.getAttribute('value')).toBe('');
+            expect(strField.getAttribute('value')).toBe('');
+            expect(dateField.getAttribute('value')).toBe('');
+            expect(listField.getAttribute('value')).toBe('');
+
+          });
+
 
           it('should merge FHIR SDC QuestionnaireResponse data back into the form', function() {
             tp.openUSSGFHTVertical();
@@ -722,10 +764,17 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
 
             var cwe = element(by.id('/type10/1'));
             var cweRepeats = element(by.id('/multiSelectCWE/1'));
-
+            // CNE field with a default value
+            var cne = element(by.id('/type9/1'));
+            // ST field with a default value
+            var st = element(by.id('/type4/1'));
             browser.wait(function() {
               return cwe.isDisplayed();
             }, tp.WAIT_TIMEOUT_1);
+
+            // the default value should not be set
+            expect(cne.getAttribute('value')).toBe('');
+            expect(st.getAttribute('value')).toBe('');
 
             expect(cwe.getAttribute('value')).toBe("user typed value");
             cwe.evaluate('item.value').then(function(val) {