+++ CHANGELOG.md
@@ -2,16 +2,25 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [19.1.0] 2019-10-24
+## [20.0.0] 2019-10-24
 ### Added
 - Added "exists" support to skip logic
-### Fixed
+### Changes
 - The trigger.code and triger.value.code were "hidden" features, but the issue of code system 
   wasn't taken into consideration and therefore was broken.
   This has been reimplemented so that when the source question is of type CNE or CWE,  
   the trigger.value will be a hash with any or all of these three fields:
   code, system, and text, and the trigger matching will be based on these fields. Please
   see the updated lforms definition for more details. 
+
+## [19.0.2] 2019-11-04
+### Fixed
+- Fixed a bug that autocompleter options are not set for items in templateOptions.
+
+## [19.0.1] 2019-10-31
+### Changed
+- Default values are no longer set when loading QuestionnaireResponse, DiagnosticReport
+  and saved form data. 
 
 ## [19.0.0] 2019-10-28
 ### Added